ML/requirements-prod.txt
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.8.0
joblib==1.5.3
holidays==0.90
sqlalchemy==2.0.46
pymysql==1.4.6
geopy==2.4.1
openmeteo-requests==1.7.5
retry-requests==2.0.0
requests==2.32.5
# All 3 tree frameworks required: champion model type varies per dish (dynamic loading via joblib)
prophet==1.1.5
xgboost==3.1.3
catboost==1.2.7
lightgbm==4.6.0
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
